+++ crate_anon/nlp_manager/regex_parser.py
@@ -28,7 +28,7 @@ crate_anon/nlp_manager/regex_parser.py
 
 """
 
-from abc import ABCMeta
+from abc import abstractmethod, ABC
 import logging
 import sys
 from typing import Any, Dict, Generator, List, Optional, TextIO, Tuple
@@ -221,7 +221,7 @@ HELP_RELATION = (
     "(e.g. '=', '<=')"
 )
 HELP_VALUE_TEXT = "Matched numerical value, as text"
-HELP_UNITS = HELP_VALUE_TEXT
+HELP_UNITS = "Matched units, as text"
 HELP_TARGET_UNIT = "Numerical value in preferred units, if known"
 HELP_TENSE_TEXT = f"Tense text, if known (e.g. '{IS}', '{WAS}')"
 HELP_TENSE = f"Calculated tense, if known (e.g. '{PAST}', '{PRESENT}')"
@@ -261,7 +261,7 @@ def common_tense(tense_text: Optional[str], relation_text: Optional[str]) \
     return tense, relation
 
 
-class NumericalResultParser(BaseNlpParser, metaclass=ABCMeta):
+class NumericalResultParser(BaseNlpParser):
     """
     DO NOT USE DIRECTLY. Base class for generic numerical results, where
     a SINGLE variable is produced.
@@ -296,7 +296,7 @@ class NumericalResultParser(BaseNlpParser, metaclass=ABCMeta):
         self.regex_str_for_debugging = regex_str_for_debugging
 
         if nlpdef is None:  # only None for debugging!
-            self.tablename = ''
+            self.tablename = self.__class__.__name__.lower()
             self.assume_preferred_unit = True
         else:
             self.tablename = nlpdef.opt_str(
@@ -347,6 +347,7 @@ class NumericalResultParser(BaseNlpParser, metaclass=ABCMeta):
             Column(FN_TENSE, String(MAX_TENSE_LENGTH), comment=HELP_TENSE),
         ]}
 
+    @abstractmethod
     def parse(self, text: str) -> Generator[Tuple[str, Dict[str, Any]],
                                             None, None]:
         # docstring in superclass
@@ -431,7 +432,7 @@ class NumericalResultParser(BaseNlpParser, metaclass=ABCMeta):
         log.info("... detailed_test: pass")
 
 
-class SimpleNumericalResultParser(NumericalResultParser, metaclass=ABCMeta):
+class SimpleNumericalResultParser(NumericalResultParser, ABC):
     """
     Base class for simple single-format numerical results. Use this when not
     only do you have a single variable to produce, but you have a single regex
@@ -593,7 +594,7 @@ class SimpleNumericalResultParser(NumericalResultParser, metaclass=ABCMeta):
             yield self.tablename, result
 
 
-class NumeratorOutOfDenominatorParser(BaseNlpParser, metaclass=ABCMeta):
+class NumeratorOutOfDenominatorParser(BaseNlpParser, ABC):
     """
     Base class for X-out-of-Y numerical results, e.g. for MMSE/ACE.
 
@@ -676,7 +677,7 @@ class NumeratorOutOfDenominatorParser(BaseNlpParser, metaclass=ABCMeta):
                          cfgsection=cfgsection,
                          commit=commit)
         if nlpdef is None:  # only None for debugging!
-            self.tablename = ''
+            self.tablename = self.__class__.__name__.lower()
         else:
             self.tablename = nlpdef.opt_str(
                 self._sectionname, 'desttable', required=True)
@@ -959,10 +960,8 @@ class ValidatorBase(BaseNlpParser):
     """
 
     def __init__(self,
-                 nlpdef: NlpDefinition,
-                 cfgsection: str,
-                 regex_str_list: List[str],
-                 validated_variable: str,
+                 nlpdef: Optional[NlpDefinition],
+                 cfgsection: Optional[str],
                  commit: bool = False) -> None:
         """
         Args:
@@ -971,6 +970,31 @@ class ValidatorBase(BaseNlpParser):
 
             cfgsection:
                 config section name in the :ref:`NLP config file <nlp_config>`
+
+            commit:
+                force a COMMIT whenever we insert data? You should specify this
+                in multiprocess mode, or you may get database deadlocks.
+        """
+        super().__init__(nlpdef=nlpdef, cfgsection=cfgsection, commit=commit)
+        validated_variable, regex_str_list = self.get_variablename_regexstrlist()  # noqa
+        self.regex_str_list = regex_str_list  # for debugging only
+        self.compiled_regex_list = [compile_regex(r) for r in regex_str_list]
+        self.variable = f"{validated_variable}_validator"
+        self.NAME = self.variable
+
+        if nlpdef is None:  # only None for debugging!
+            self.tablename = self.__class__.__name__.lower()
+        else:
+            self.tablename = nlpdef.opt_str(
+                self._sectionname, 'desttable', required=True)
+
+    @abstractmethod
+    def get_variablename_regexstrlist(self) -> Tuple[str, List[str]]:
+        """
+        To be overridden.
+
+        Returns:
+            tuple: ``(validated_variable_name, regex_str_list)``, where:
 
             regex_str_list:
                 List of regular expressions, each in string format.
@@ -986,21 +1010,8 @@ class ValidatorBase(BaseNlpParser):
                 ``validated_variable == 'crp'``, then the ``variable_name``
                 field will be set to ``crp_validator``.
 
-            commit:
-                force a COMMIT whenever we insert data? You should specify this
-                in multiprocess mode, or you may get database deadlocks.
         """
-        super().__init__(nlpdef=nlpdef, cfgsection=cfgsection, commit=commit)
-        self.regex_str_list = regex_str_list  # for debugging only
-        self.compiled_regex_list = [compile_regex(r) for r in regex_str_list]
-        self.variable = f"{validated_variable}_validator"
-        self.NAME = self.variable
-
-        if nlpdef is None:  # only None for debugging!
-            self.tablename = ''
-        else:
-            self.tablename = nlpdef.opt_str(
-                self._sectionname, 'desttable', required=True)
+        raise NotImplementedError
 
     def print_info(self, file: TextIO = sys.stdout) -> None:
         # docstring in superclass